pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "modelmesh"
version = "0.2.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    "httpx>=0.27.0",
    "openai>=1.30.0",
    "anthropic>=0.28.0",
    "PyYAML>=6.0.1",
    "python-json-logger>=2.0.7",
    # Phase 2
    "prometheus-client>=0.20.0",
    "passlib[bcrypt]>=1.7.4",
    "redis[asyncio]>=5.0.0",
]

[project.optional-dependencies]
semantic = [
    "sentence-transformers>=3.0.0",
]
huggingface = [
    "huggingface-hub>=0.23.0",
]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-httpx>=0.30.0",
    "httpx>=0.27.0",
    "respx>=0.21.0",
    "fakeredis[aioredis]>=2.23.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
